+++ toolkit/components/alerts/AlertNotificationIPCSerializer.h
@@ -29,8 +29,9 @@ struct IPDLParamTraits<nsIAlertNotification*> {
     }
 
     nsString name, imageURL, title, text, cookie, dir, lang, data;
-    bool textClickable, inPrivateBrowsing, requireInteraction;
+    bool textClickable, inPrivateBrowsing, requireInteraction, silent;
     nsCOMPtr<nsIPrincipal> principal;
+    nsTArray<uint32_t> vibrate;
 
     if (NS_WARN_IF(NS_FAILED(aParam->GetName(name))) ||
         NS_WARN_IF(NS_FAILED(aParam->GetImageURL(imageURL))) ||
@@ -46,7 +47,11 @@ struct IPDLParamTraits<nsIAlertNotification*> {
         NS_WARN_IF(
             NS_FAILED(aParam->GetInPrivateBrowsing(&inPrivateBrowsing))) ||
         NS_WARN_IF(
-            NS_FAILED(aParam->GetRequireInteraction(&requireInteraction)))) {
+            NS_FAILED(aParam->GetRequireInteraction(&requireInteraction))) ||
+        NS_WARN_IF(
+            NS_FAILED(aParam->GetSilent(&silent))) ||
+        NS_WARN_IF(
+            NS_FAILED(aParam->GetVibrate(vibrate)))) {
       // Write a `null` object if any getter returns an error. Otherwise, the
       // receiver will try to deserialize an incomplete object and crash.
       WriteIPDLParam(aMsg, aActor, /* isNull */ true);
@@ -66,6 +71,8 @@ struct IPDLParamTraits<nsIAlertNotification*> {
     WriteIPDLParam(aMsg, aActor, IPC::Principal(principal));
     WriteIPDLParam(aMsg, aActor, inPrivateBrowsing);
     WriteIPDLParam(aMsg, aActor, requireInteraction);
+    WriteIPDLParam(aMsg, aActor, silent);
+    WriteIPDLParam(aMsg, aActor, vibrate);
   }
 
   static bool Read(const IPC::Message* aMsg, PickleIterator* aIter,
@@ -78,8 +85,9 @@ struct IPDLParamTraits<nsIAlertNotification*> {
     }
 
     nsString name, imageURL, title, text, cookie, dir, lang, data;
-    bool textClickable, inPrivateBrowsing, requireInteraction;
+    bool textClickable, inPrivateBrowsing, requireInteraction, silent;
     IPC::Principal principal;
+    nsTArray<uint32_t> vibrate;
 
     if (!ReadIPDLParam(aMsg, aIter, aActor, &name) ||
         !ReadIPDLParam(aMsg, aIter, aActor, &imageURL) ||
@@ -92,7 +100,9 @@ struct IPDLParamTraits<nsIAlertNotification*> {
         !ReadIPDLParam(aMsg, aIter, aActor, &data) ||
         !ReadIPDLParam(aMsg, aIter, aActor, &principal) ||
         !ReadIPDLParam(aMsg, aIter, aActor, &inPrivateBrowsing) ||
-        !ReadIPDLParam(aMsg, aIter, aActor, &requireInteraction)) {
+        !ReadIPDLParam(aMsg, aIter, aActor, &requireInteraction) ||
+        !ReadIPDLParam(aMsg, aIter, aActor, &silent) ||
+        !ReadIPDLParam(aMsg, aIter, aActor, &vibrate)) {
       return false;
     }
 
@@ -104,7 +114,8 @@ struct IPDLParamTraits<nsIAlertNotification*> {
     }
     nsresult rv = alert->Init(name, imageURL, title, text, textClickable,
                               cookie, dir, lang, data, principal,
-                              inPrivateBrowsing, requireInteraction);
+                              inPrivateBrowsing, requireInteraction,
+                              silent, vibrate);
     if (NS_WARN_IF(NS_FAILED(rv))) {
       *aResult = nullptr;
       return true;