+++ toolkit/system/gnome/nsSystemAlertsService.cpp
@@ -33,10 +33,13 @@ NS_IMETHODIMP nsSystemAlertsService::ShowAlertNotification(
   nsCOMPtr<nsIAlertNotification> alert =
       do_CreateInstance(ALERT_NOTIFICATION_CONTRACTID);
   NS_ENSURE_TRUE(alert, NS_ERROR_FAILURE);
+  // vibrate is unused for now
+  nsTArray<uint32_t> vibrate;
   nsresult rv =
       alert->Init(aAlertName, aImageUrl, aAlertTitle, aAlertText,
                   aAlertTextClickable, aAlertCookie, aBidi, aLang, aData,
-                  aPrincipal, aInPrivateBrowsing, aRequireInteraction);
+                  aPrincipal, aInPrivateBrowsing, aRequireInteraction,
+                  false, vibrate);
   NS_ENSURE_SUCCESS(rv, rv);
   return ShowAlert(alert, aAlertListener);
 }