+++ widget/cocoa/OSXNotificationCenter.mm
@@ -203,9 +203,12 @@ OSXNotificationCenter::ShowAlertNotification(
     bool aRequireInteraction) {
   nsCOMPtr<nsIAlertNotification> alert = do_CreateInstance(ALERT_NOTIFICATION_CONTRACTID);
   NS_ENSURE_TRUE(alert, NS_ERROR_FAILURE);
+  // vibrate is unused for now
+  nsTArray<uint32_t> vibrate;
   nsresult rv =
       alert->Init(aAlertName, aImageUrl, aAlertTitle, aAlertText, aAlertTextClickable, aAlertCookie,
-                  aBidi, aLang, aData, aPrincipal, aInPrivateBrowsing, aRequireInteraction);
+                  aBidi, aLang, aData, aPrincipal, aInPrivateBrowsing, aRequireInteraction,
+                  false, vibrate);
   NS_ENSURE_SUCCESS(rv, rv);
   return ShowAlert(alert, aAlertListener);
 }