+++ widget/windows/ToastNotification.cpp
@@ -86,10 +86,13 @@ ToastNotification::ShowAlertNotification(
   if (NS_WARN_IF(!alert)) {
     return NS_ERROR_FAILURE;
   }
+  // vibrate is unused for now
+  nsTArray<uint32_t> vibrate;
   nsresult rv =
       alert->Init(aAlertName, aImageUrl, aAlertTitle, aAlertText,
                   aAlertTextClickable, aAlertCookie, aBidi, aLang, aData,
-                  aPrincipal, aInPrivateBrowsing, aRequireInteraction);
+                  aPrincipal, aInPrivateBrowsing, aRequireInteraction,
+                  false, vibrate);
   if (NS_WARN_IF(NS_FAILED(rv))) {
     return rv;
   }